yiyu Toolbox core/requirements.txt
torch
torchvision
opencv-python
scenedetect[opencv]
PyQt5
numpy
rembg
Pillow
onnxruntime
